backend/etl-enrich/requirements.txt
nats-py==2.6.0
psycopg[binary]==3.1.18
psycopg_pool==3.2.3
neo4j==5.15.0
pydantic==2.5.2
orjson==3.9.10
python-dotenv==1.0.0
tenacity==8.2.3
pytest==7.4.3
pytest-asyncio==0.21.1
